Starter for the ironman mode guide support.
Does not detect the account type but instead requires you to toggle the
guide to the ironman mode via the config.

The quests_ironman.json contains a duplicate Romeo & Juliet quest that does
not get marked as completed. Need to see why this is a duplicate on the
wiki in the first place. The quest Sins of the Father is also not listed
within the guide.

The quest requirements are still taken from the Skill Requirements page
of the wiki and are not inline with the requirements listed on the
Ironman optimal quest guide page. This will require an update to
oqg-parser to support parsing the 2nd column of requirements.

See issue #12